Remove ToObject from ICustomObject, fix negative tests, add new tests
